Deep_Learning_ImageCaptioning/requirements.txt
numpy==1.18.1
numpydoc==0.9.2
Pillow==6.2.2
pprint==0.1
PyAutoGUI==0.9.48
pycocoevalcap==1.1
pycocotools==2.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.4.1
sklearn==0.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.39.0
